Создан utils/render.js
